+++ examples/file_and_images/android/app/src/main/kotlin/es/antonborri/file_and_images/MainActivity.kt
@@ -1,0 +1,5 @@
+package es.antonborri.file_and_images
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()